+++ test/instabug_flutter_test.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 import 'dart:typed_data';
 import 'dart:ui';
 
+import 'package:flutter/cupertino.dart';
 import 'package:instabug_flutter/Instabug.dart';
 import 'package:instabug_flutter/BugReporting.dart';
 import 'package:instabug_flutter/InstabugLog.dart';
@@ -13,15 +14,19 @@ import 'package:instabug_flutter/Chats.dart';
 import 'package:instabug_flutter/Replies.dart';
 
 void main() {
-
+  WidgetsFlutterBinding.ensureInitialized();
   final List<MethodCall> log = <MethodCall>[];
   final appToken = '068ba9a8c3615035e163dc5f829c73be';
-  final List<InvocationEvent> invocationEvents = <InvocationEvent>[InvocationEvent.floatingButton];
+  final List<InvocationEvent> invocationEvents = <InvocationEvent>[
+    InvocationEvent.floatingButton
+  ];
   final email = "s@nta.com";
   final name = "santa";
   String message = "Test Message";
   const String userAttribute = '19';
-  const Map<String, String> userAttributePair = <String, String>{'gender': 'female'};
+  const Map<String, String> userAttributePair = <String, String>{
+    'gender': 'female'
+  };
 
   setUpAll(() async {
     MethodChannel('instabug_flutter')
@@ -29,7 +34,7 @@ void main() {
       log.add(methodCall);
       switch (methodCall.method) {
         case 'getTags':
-          return <String>['tag1', 'tag2']; 
+          return <String>['tag1', 'tag2'];
         case 'getUserAttributeForKey:':
           return userAttribute;
         case 'getUserAttributes':
@@ -41,24 +46,29 @@ void main() {
   });
 
   tearDown(() async {
-      log.clear();
-    });
+    log.clear();
+  });
 
-test('startWithToken:invocationEvents: Test', () async {
+  test('startWithToken:invocationEvents: Test', () async {
     Instabug.start(appToken, invocationEvents);
-    final List<dynamic> args = <dynamic>[appToken, <String>[InvocationEvent.floatingButton.toString()]];
+    final List<dynamic> args = <dynamic>[
+      appToken,
+      <String>[InvocationEvent.floatingButton.toString()]
+    ];
     expect(log, <Matcher>[
-      isMethodCall('startWithToken:invocationEvents:',
+      isMethodCall(
+        'startWithToken:invocationEvents:',
         arguments: args,
       )
     ]);
   });
-  
+
   test('showWelcomeMessageWithMode: Test', () async {
     Instabug.showWelcomeMessageWithMode(WelcomeMessageMode.beta);
     final List<dynamic> args = <dynamic>[WelcomeMessageMode.beta.toString()];
     expect(log, <Matcher>[
-      isMethodCall('showWelcomeMessageWithMode:',
+      isMethodCall(
+        'showWelcomeMessageWithMode:',
         arguments: args,
       )
     ]);
@@ -68,66 +78,68 @@ test('startWithToken:invocationEvents: Test', () async {
     Instabug.identifyUser(email, name);
     final List<dynamic> args = <dynamic>[email, name];
     expect(log, <Matcher>[
-      isMethodCall('identifyUserWithEmail:name:',
+      isMethodCall(
+        'identifyUserWithEmail:name:',
         arguments: args,
       )
     ]);
   });
-  
+
   test('identifyUserWithEmail:name: Test Optional Parameter', () async {
     Instabug.identifyUser(email);
     final List<dynamic> args = <dynamic>[email, null];
     expect(log, <Matcher>[
-      isMethodCall('identifyUserWithEmail:name:',
-        arguments:args,
+      isMethodCall(
+        'identifyUserWithEmail:name:',
+        arguments: args,
       )
     ]);
   });
 
   test('logOut Test', () async {
     Instabug.logOut();
-    expect(log, <Matcher>[
-      isMethodCall('logOut',
-        arguments: null)
-    ]);
+    expect(log, <Matcher>[isMethodCall('logOut', arguments: null)]);
   });
-  
+
   test('setLocale:', () async {
     Instabug.setLocale(IBGLocale.german);
     final List<dynamic> args = <dynamic>[IBGLocale.german.toString()];
     expect(log, <Matcher>[
-      isMethodCall('setLocale:',
+      isMethodCall(
+        'setLocale:',
         arguments: args,
       )
     ]);
   });
-
 
   test('logVerbose: Test', () async {
     InstabugLog.logVerbose(message);
     final List<dynamic> args = <dynamic>[message];
     expect(log, <Matcher>[
-      isMethodCall('logVerbose:',
+      isMethodCall(
+        'logVerbose:',
         arguments: args,
       )
     ]);
   });
-  
+
   test('logDebug: Test', () async {
     InstabugLog.logDebug(message);
     final List<dynamic> args = <dynamic>[message];
     expect(log, <Matcher>[
-      isMethodCall('logDebug:',
+      isMethodCall(
+        'logDebug:',
         arguments: args,
       )
     ]);
   });
-  
+
   test('logInfo: Test', () async {
     InstabugLog.logInfo(message);
     final List<dynamic> args = <dynamic>[message];
     expect(log, <Matcher>[
-      isMethodCall('logInfo:',
+      isMethodCall(
+        'logInfo:',
         arguments: args,
       )
     ]);
@@ -135,50 +147,52 @@ test('startWithToken:invocationEvents: Test', () async {
 
   test('clearAllLogs: Test', () async {
     InstabugLog.clearAllLogs();
-    expect(log, <Matcher>[
-      isMethodCall('clearAllLogs',
-        arguments: null
-      )
-    ]);
+    expect(log, <Matcher>[isMethodCall('clearAllLogs', arguments: null)]);
   });
 
   test('logError: Test', () async {
     InstabugLog.logError(message);
     final List<dynamic> args = <dynamic>[message];
     expect(log, <Matcher>[
-      isMethodCall('logError:',
+      isMethodCall(
+        'logError:',
         arguments: args,
       )
     ]);
   });
-  
+
   test('logWarn: Test', () async {
     InstabugLog.logWarn(message);
     final List<dynamic> args = <dynamic>[message];
     expect(log, <Matcher>[
-      isMethodCall('logWarn:',
+      isMethodCall(
+        'logWarn:',
         arguments: args,
       )
     ]);
   });
 
-  test('test setColorTheme should be called with argument colorTheme', () async {
+  test('test setColorTheme should be called with argument colorTheme',
+      () async {
     const ColorTheme colorTheme = ColorTheme.dark;
     Instabug.setColorTheme(colorTheme);
     final List<dynamic> args = <dynamic>[colorTheme.toString()];
     expect(log, <Matcher>[
-      isMethodCall('setColorTheme:',
+      isMethodCall(
+        'setColorTheme:',
         arguments: args,
       )
     ]);
   });
 
-  test('test appendTags should be called with argument List of strings', () async {
+  test('test appendTags should be called with argument List of strings',
+      () async {
     const List<String> tags = ['tag1', 'tag2'];
     Instabug.appendTags(tags);
     final List<dynamic> args = <dynamic>[tags];
     expect(log, <Matcher>[
-      isMethodCall('appendTags:',
+      isMethodCall(
+        'appendTags:',
         arguments: args,
       )
     ]);
@@ -186,71 +200,69 @@ test('startWithToken:invocationEvents: Test', () async {
 
   test('test resetTags should be called with no arguments', () async {
     Instabug.resetTags();
-    expect(log, <Matcher>[
-      isMethodCall('resetTags',
-        arguments: null
-      )
-    ]);
+    expect(log, <Matcher>[isMethodCall('resetTags', arguments: null)]);
   });
 
-  test('test getTags should be called with no arguments and returns list of tags', () async {
+  test(
+      'test getTags should be called with no arguments and returns list of tags',
+      () async {
     List<String> tags = await Instabug.getTags();
-    expect(log, <Matcher>[
-      isMethodCall('getTags',
-        arguments: null
-      )
-    ]);
-    expect(tags, ['tag1','tag2']);
+    expect(log, <Matcher>[isMethodCall('getTags', arguments: null)]);
+    expect(tags, ['tag1', 'tag2']);
   });
 
-  test('test setUserAttributeWithKey should be called with two string arguments', () async {
+  test(
+      'test setUserAttributeWithKey should be called with two string arguments',
+      () async {
     const String value = '19';
     const String key = 'Age';
     Instabug.setUserAttribute(value, key);
     final List<dynamic> args = <dynamic>[value, key];
     expect(log, <Matcher>[
-      isMethodCall('setUserAttribute:withKey:',
+      isMethodCall(
+        'setUserAttribute:withKey:',
         arguments: args,
       )
     ]);
   });
 
-  test('test removeUserAttributeForKey should be called with a string argument', () async {
+  test('test removeUserAttributeForKey should be called with a string argument',
+      () async {
     const String key = 'Age';
     Instabug.removeUserAttribute(key);
     final List<dynamic> args = <dynamic>[key];
     expect(log, <Matcher>[
-      isMethodCall('removeUserAttributeForKey:',
+      isMethodCall(
+        'removeUserAttributeForKey:',
         arguments: args,
       )
     ]);
   });
 
-  test('test getUserAttributeForKey should be called with a string argument and return a string', () async {
+  test(
+      'test getUserAttributeForKey should be called with a string argument and return a string',
+      () async {
     const String key = 'Age';
     final String value = await Instabug.getUserAttributeForKey(key);
     expect(log, <Matcher>[
-      isMethodCall('getUserAttributeForKey:',
-        arguments: <dynamic>[key]
-      )
+      isMethodCall('getUserAttributeForKey:', arguments: <dynamic>[key])
     ]);
     expect(value, userAttribute);
   });
 
-  test('test getuserAttributes should be called with no arguments and returns a Map', () async {
+  test(
+      'test getuserAttributes should be called with no arguments and returns a Map',
+      () async {
     final Map<String, String> result = await Instabug.getUserAttributes();
-    expect(log, <Matcher>[
-      isMethodCall('getUserAttributes',
-        arguments: null
-      )
-    ]);
+    expect(log, <Matcher>[isMethodCall('getUserAttributes', arguments: null)]);
     expect(result, userAttributePair);
   });
-  
+
   test('show Test', () async {
     Instabug.show();
     expect(log, <Matcher>[
-      isMethodCall('show',
+      isMethodCall(
+        'show',
         arguments: null,
       )
     ]);
@@ -260,19 +272,22 @@ test('startWithToken:invocationEvents: Test', () async {
     Instabug.logUserEvent(name);
     final List<dynamic> args = <dynamic>[name];
     expect(log, <Matcher>[
-      isMethodCall('logUserEventWithName:',
+      isMethodCall(
+        'logUserEventWithName:',
         arguments: args,
       )
     ]);
   });
 
-  test('test setValueForStringWithKey should be called with two arguments', () async {
+  test('test setValueForStringWithKey should be called with two arguments',
+      () async {
     const String value = 'Some key';
     const CustomTextPlaceHolderKey key = CustomTextPlaceHolderKey.shakeHint;
     Instabug.setValueForStringWithKey(value, key);
     final List<dynamic> args = <dynamic>[value, key.toString()];
     expect(log, <Matcher>[
-      isMethodCall('setValue:forStringWithKey:',
+      isMethodCall(
+        'setValue:forStringWithKey:',
         arguments: args,
       )
     ]);
@@ -283,7 +298,8 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[sessionProfilerEnabled];
     Instabug.setSessionProfilerEnabled(sessionProfilerEnabled);
     expect(log, <Matcher>[
-      isMethodCall('setSessionProfilerEnabled:',
+      isMethodCall(
+        'setSessionProfilerEnabled:',
         arguments: args,
       )
     ]);
@@ -294,7 +310,8 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[c.value];
     Instabug.setPrimaryColor(c);
     expect(log, <Matcher>[
-      isMethodCall('setPrimaryColor:',
+      isMethodCall(
+        'setPrimaryColor:',
         arguments: args,
       )
     ]);
@@ -305,7 +322,8 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[s];
     Instabug.setUserData(s);
     expect(log, <Matcher>[
-      isMethodCall('setUserData:',
+      isMethodCall(
+        'setUserData:',
         arguments: args,
       )
     ]);
@@ -315,9 +333,10 @@ test('startWithToken:invocationEvents: Test', () async {
     String filePath = "filePath";
     String fileName = "fileName";
     final List<dynamic> args = <dynamic>[filePath, fileName];
-    Instabug.addFileAttachmentWithURL(filePath,fileName);
+    Instabug.addFileAttachmentWithURL(filePath, fileName);
     expect(log, <Matcher>[
-      isMethodCall('addFileAttachmentWithURL:',
+      isMethodCall(
+        'addFileAttachmentWithURL:',
         arguments: args,
       )
     ]);
@@ -327,9 +346,10 @@ test('startWithToken:invocationEvents: Test', () async {
     var bdata = new Uint8List(10);
     String fileName = "fileName";
     final List<dynamic> args = <dynamic>[bdata, fileName];
-    Instabug.addFileAttachmentWithData(bdata,fileName);
+    Instabug.addFileAttachmentWithData(bdata, fileName);
     expect(log, <Matcher>[
-      isMethodCall('addFileAttachmentWithData:',
+      isMethodCall(
+        'addFileAttachmentWithData:',
         arguments: args,
       )
     ]);
@@ -338,7 +358,8 @@ test('startWithToken:invocationEvents: Test', () async {
   test('clearFileAttachments Test', () async {
     Instabug.clearFileAttachments();
     expect(log, <Matcher>[
-      isMethodCall('clearFileAttachments',
+      isMethodCall(
+        'clearFileAttachments',
         arguments: null,
       )
     ]);
@@ -348,7 +369,8 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[WelcomeMessageMode.live.toString()];
     Instabug.setWelcomeMessageMode(WelcomeMessageMode.live);
     expect(log, <Matcher>[
-      isMethodCall('setWelcomeMessageMode:',
+      isMethodCall(
+        'setWelcomeMessageMode:',
         arguments: args,
       )
     ]);
@@ -359,85 +381,111 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[isEnabled];
     BugReporting.setEnabled(isEnabled);
     expect(log, <Matcher>[
-      isMethodCall('setBugReportingEnabled:',
+      isMethodCall(
+        'setBugReportingEnabled:',
         arguments: args,
       )
     ]);
   });
 
   test('setOnInvokeCallback Test', () async {
-    BugReporting.setOnInvokeCallback(()=> (){});
+    BugReporting.setOnInvokeCallback(() => () {});
     expect(log, <Matcher>[
-      isMethodCall('setOnInvokeCallback',
+      isMethodCall(
+        'setOnInvokeCallback',
         arguments: null,
       )
     ]);
   });
 
   test('setOnDismissCallback Test', () async {
-    BugReporting.setOnDismissCallback(()=> (){});
+    BugReporting.setOnDismissCallback(() => () {});
     expect(log, <Matcher>[
-      isMethodCall('setOnDismissCallback',
+      isMethodCall(
+        'setOnDismissCallback',
         arguments: null,
       )
     ]);
   });
 
   test('setInvocationEvents Test', () async {
-     BugReporting.setInvocationEvents(<InvocationEvent>[InvocationEvent.floatingButton]);
-    final List<dynamic> args = <dynamic>[<String>[InvocationEvent.floatingButton.toString()]];
+    BugReporting.setInvocationEvents(
+        <InvocationEvent>[InvocationEvent.floatingButton]);
+    final List<dynamic> args = <dynamic>[
+      <String>[InvocationEvent.floatingButton.toString()]
+    ];
     expect(log, <Matcher>[
-      isMethodCall('setInvocationEvents:',
+      isMethodCall(
+        'setInvocationEvents:',
         arguments: args,
       )
     ]);
   });
 
-  test('setEnabledAttachmentTypes:extraScreenShot:galleryImage:screenRecording: Test', () async {
+  test(
+      'setEnabledAttachmentTypes:extraScreenShot:galleryImage:screenRecording: Test',
+      () async {
     BugReporting.setEnabledAttachmentTypes(false, false, false, false);
     final List<dynamic> args = <dynamic>[false, false, false, false];
     expect(log, <Matcher>[
-      isMethodCall('setEnabledAttachmentTypes:extraScreenShot:galleryImage:screenRecording:',
-        arguments: args,
-      )
-    ]);
-  });
-
- test('setInvocationEvents Test', () async {
-    BugReporting.setReportTypes(<ReportType>[ReportType.feedback]);
-    final List<dynamic> args = <dynamic>[<String>[ReportType.feedback.toString()]];
-    expect(log, <Matcher>[
-      isMethodCall('setReportTypes:',
+      isMethodCall(
+        'setEnabledAttachmentTypes:extraScreenShot:galleryImage:screenRecording:',
         arguments: args,
       )
     ]);
   });
 
   test('setInvocationEvents Test', () async {
-    BugReporting.setExtendedBugReportMode(ExtendedBugReportMode.enabledWithOptionalFields);
-    final List<dynamic> args = <dynamic>[ExtendedBugReportMode.enabledWithOptionalFields.toString()];
+    BugReporting.setReportTypes(<ReportType>[ReportType.feedback]);
+    final List<dynamic> args = <dynamic>[
+      <String>[ReportType.feedback.toString()]
+    ];
     expect(log, <Matcher>[
-      isMethodCall('setExtendedBugReportMode:',
+      isMethodCall(
+        'setReportTypes:',
+        arguments: args,
+      )
+    ]);
+  });
+
+  test('setInvocationEvents Test', () async {
+    BugReporting.setExtendedBugReportMode(
+        ExtendedBugReportMode.enabledWithOptionalFields);
+    final List<dynamic> args = <dynamic>[
+      ExtendedBugReportMode.enabledWithOptionalFields.toString()
+    ];
+    expect(log, <Matcher>[
+      isMethodCall(
+        'setExtendedBugReportMode:',
         arguments: args,
       )
     ]);
   });
 
   test('setInvocationOptions Test', () async {
-    BugReporting.setInvocationOptions(<InvocationOption>[InvocationOption.emailFieldHidden]);
-    final List<dynamic> args = <dynamic>[<String>[InvocationOption.emailFieldHidden.toString()]];
+    BugReporting.setInvocationOptions(
+        <InvocationOption>[InvocationOption.emailFieldHidden]);
+    final List<dynamic> args = <dynamic>[
+      <String>[InvocationOption.emailFieldHidden.toString()]
+    ];
     expect(log, <Matcher>[
-      isMethodCall('setInvocationOptions:',
+      isMethodCall(
+        'setInvocationOptions:',
         arguments: args,
       )
     ]);
   });
 
-   test('showBugReportingWithReportTypeAndOptions:options Test', () async {
-   BugReporting.show(ReportType.bug, <InvocationOption>[InvocationOption.emailFieldHidden]);
-    final List<dynamic> args = <dynamic>[ReportType.bug.toString(), <String>[InvocationOption.emailFieldHidden.toString()]];
+  test('showBugReportingWithReportTypeAndOptions:options Test', () async {
+    BugReporting.show(
+        ReportType.bug, <InvocationOption>[InvocationOption.emailFieldHidden]);
+    final List<dynamic> args = <dynamic>[
+      ReportType.bug.toString(),
+      <String>[InvocationOption.emailFieldHidden.toString()]
+    ];
     expect(log, <Matcher>[
-      isMethodCall('showBugReportingWithReportTypeAndOptions:options:',
+      isMethodCall(
+        'showBugReportingWithReportTypeAndOptions:options:',
         arguments: args,
       )
     ]);
@@ -448,7 +496,8 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[isEnabled];
     Surveys.setEnabled(isEnabled);
     expect(log, <Matcher>[
-      isMethodCall('setSurveysEnabled:',
+      isMethodCall(
+        'setSurveysEnabled:',
         arguments: args,
       )
     ]);
@@ -459,25 +508,28 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[isEnabled];
     Surveys.setAutoShowingEnabled(isEnabled);
     expect(log, <Matcher>[
-      isMethodCall('setAutoShowingSurveysEnabled:',
+      isMethodCall(
+        'setAutoShowingSurveysEnabled:',
         arguments: args,
       )
     ]);
   });
 
   test('setOnShowSurveyCallback Test', () async {
-    Surveys.setOnShowCallback(()=> (){});
+    Surveys.setOnShowCallback(() => () {});
     expect(log, <Matcher>[
-      isMethodCall('setOnShowSurveyCallback',
+      isMethodCall(
+        'setOnShowSurveyCallback',
         arguments: null,
       )
     ]);
   });
 
   test('setOnDismissSurveyCallback Test', () async {
-    Surveys.setOnDismissCallback(()=> (){});
+    Surveys.setOnDismissCallback(() => () {});
     expect(log, <Matcher>[
-      isMethodCall('setOnDismissSurveyCallback',
+      isMethodCall(
+        'setOnDismissSurveyCallback',
         arguments: null,
       )
     ]);
@@ -488,7 +540,8 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[isEnabled];
     Surveys.setShouldShowWelcomeScreen(isEnabled);
     expect(log, <Matcher>[
-      isMethodCall('setShouldShowSurveysWelcomeScreen:',
+      isMethodCall(
+        'setShouldShowSurveysWelcomeScreen:',
         arguments: args,
       )
     ]);
@@ -497,7 +550,8 @@ test('startWithToken:invocationEvents: Test', () async {
   test('showSurveysIfAvailable Test', () async {
     Surveys.showSurveyIfAvailable();
     expect(log, <Matcher>[
-      isMethodCall('showSurveysIfAvailable',
+      isMethodCall(
+        'showSurveysIfAvailable',
       )
     ]);
   });
@@ -507,8 +561,9 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[token];
     Surveys.showSurvey(token);
     expect(log, <Matcher>[
-      isMethodCall('showSurveyWithToken:',
-      arguments: args,
+      isMethodCall(
+        'showSurveyWithToken:',
+        arguments: args,
       )
     ]);
   });
@@ -516,39 +571,39 @@ test('startWithToken:invocationEvents: Test', () async {
   test('hasRespondedToSurvey Test', () async {
     String token = "token";
     final List<dynamic> args = <dynamic>[token];
-    Surveys.hasRespondedToSurvey(token,()=> (){});
+    Surveys.hasRespondedToSurvey(token, () => () {});
     expect(log, <Matcher>[
-      isMethodCall('hasRespondedToSurveyWithToken:',
-      arguments: args,
+      isMethodCall(
+        'hasRespondedToSurveyWithToken:',
+        arguments: args,
       )
     ]);
   });
 
   test('showFeatureRequests Test', () async {
     FeatureRequests.show();
-    expect(log, <Matcher>[
-      isMethodCall('showFeatureRequests'
-     )
-    ]);
+    expect(log, <Matcher>[isMethodCall('showFeatureRequests')]);
   });
 
   test('setEmailFieldRequiredForFeatureRequests:forAction: Test', () async {
     bool isEmailFieldRequired = false;
-    final List<dynamic> args = <dynamic>[isEmailFieldRequired, <String>[ActionType.allActions.toString()]];
-    FeatureRequests.setEmailFieldRequired(isEmailFieldRequired, [ActionType.allActions]);
+    final List<dynamic> args = <dynamic>[
+      isEmailFieldRequired,
+      <String>[ActionType.addCommentToFeature.toString()]
+    ];
+    FeatureRequests.setEmailFieldRequired(
+        isEmailFieldRequired, [ActionType.addCommentToFeature]);
     expect(log, <Matcher>[
-      isMethodCall('setEmailFieldRequiredForFeatureRequests:forAction:',
-      arguments: args,
-     )
+      isMethodCall(
+        'setEmailFieldRequiredForFeatureRequests:forAction:',
+        arguments: args,
+      )
     ]);
   });
 
   test('showChats Test', () async {
     Chats.show();
-    expect(log, <Matcher>[
-      isMethodCall('showChats'
-     )
-    ]);
+    expect(log, <Matcher>[isMethodCall('showChats')]);
   });
 
   test('setChatsEnabled: Test', () async {
@@ -556,7 +611,8 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[isEnabled];
     Chats.setEnabled(isEnabled);
     expect(log, <Matcher>[
-      isMethodCall('setChatsEnabled:',
+      isMethodCall(
+        'setChatsEnabled:',
         arguments: args,
       )
     ]);
@@ -567,7 +623,8 @@ test('startWithToken:invocationEvents: Test', () async {
     final List<dynamic> args = <dynamic>[isEnabled];
     Replies.setEnabled(isEnabled);
     expect(log, <Matcher>[
-      isMethodCall('setRepliesEnabled:',
+      isMethodCall(
+        'setRepliesEnabled:',
         arguments: args,
       )
     ]);
@@ -575,61 +632,46 @@ test('startWithToken:invocationEvents: Test', () async {
 
   test('showReplies Test', () async {
     Replies.show();
-    expect(log, <Matcher>[
-      isMethodCall('showReplies'
-     )
-    ]);
+    expect(log, <Matcher>[isMethodCall('showReplies')]);
   });
 
   test('hasChats Test', () async {
-    Replies.hasChats(()=> (){});
-    expect(log, <Matcher>[
-      isMethodCall('hasChats'
-      )
-    ]);
+    Replies.hasChats(() => () {});
+    expect(log, <Matcher>[isMethodCall('hasChats')]);
   });
 
   test('setOnNewReplyReceivedCallback Test', () async {
-    Replies.setOnNewReplyReceivedCallback(()=> (){});
-    expect(log, <Matcher>[
-      isMethodCall('setOnNewReplyReceivedCallback'
-      )
-    ]);
+    Replies.setOnNewReplyReceivedCallback(() => () {});
+    expect(log, <Matcher>[isMethodCall('setOnNewReplyReceivedCallback')]);
   });
 
   test('getUnreadRepliesCount Test', () async {
-    Replies.getUnreadRepliesCount(()=> (){});
-    expect(log, <Matcher>[
-      isMethodCall('getUnreadRepliesCount'
-      )
-    ]);
+    Replies.getUnreadRepliesCount(() => () {});
+    expect(log, <Matcher>[isMethodCall('getUnreadRepliesCount')]);
   });
 
-   test('setChatNotificationEnabled: Test', () async {
+  test('setChatNotificationEnabled: Test', () async {
     bool isEnabled = false;
     final List<dynamic> args = <dynamic>[isEnabled];
     Replies.setInAppNotificationsEnabled(isEnabled);
     expect(log, <Matcher>[
-      isMethodCall('setChatNotificationEnabled:',
+      isMethodCall(
+        'setChatNotificationEnabled:',
         arguments: args,
       )
     ]);
   });
 
-
   ///Since the below method only runs on android and has the [Platform.isAndroid] condition in it, it will fail when running outside android,
   /// therefore its commented.
-    // test('setEnableInAppNotificationSound: Test', () async {
-    //   bool isEnabled = false;
-    //   final List<dynamic> args = <dynamic>[isEnabled];
-    //   Replies.setInAppNotificationSound(isEnabled);
-    //   expect(log, <Matcher>[
-    //     isMethodCall('setEnableInAppNotificationSound:',
-    //       arguments: args,
-    //     )
-    //   ]);
-    // });
-
+  // test('setEnableInAppNotificationSound: Test', () async {
+  //   bool isEnabled = false;
+  //   final List<dynamic> args = <dynamic>[isEnabled];
+  //   Replies.setInAppNotificationSound(isEnabled);
+  //   expect(log, <Matcher>[
+  //     isMethodCall('setEnableInAppNotificationSound:',
+  //       arguments: args,
+  //     )
+  //   ]);
+  // });
 }
-
-